Moved more elements in the SwiftUI HeaderView
