+++ brainwallet/New Lock Screen Classes/LockScreenViewModel.swift
