+++ brainwallet/Storyboards/Main.storyboard
@@ -27,30 +27,37 @@
                             <view contentMode="scaleToFill" insetsLayoutMarginsFromSafeArea="NO" translatesAutoresizingMaskIntoConstraints="NO" id="8fw-XB-zyN">
                                 <rect key="frame" x="0.0" y="20" width="414" height="80"/>
                                 <subviews>
-                                    <button opaque="NO" contentMode="scaleToFill" fixedFrame="YES" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="jLr-H0-dij">
-                                        <rect key="frame" x="380" y="18" width="24" height="24"/>
-                                        <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                                    <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="jLr-H0-dij">
+                                        <rect key="frame" x="358" y="12" width="44" height="44"/>
+                                        <constraints>
+                                            <constraint firstAttribute="width" constant="44" id="Sym-3S-iFL"/>
+                                            <constraint firstAttribute="height" constant="44" id="Ykw-TZ-IhF"/>
+                                        </constraints>
                                         <color key="tintColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                                        <state key="normal" image="settings_hamburger"/>
+                                        <state key="normal" image="settings_hamburger">
+                                            <color key="titleColor" name="brainwalletContent"/>
+                                        </state>
                                         <connections>
                                             <action selector="showSettingsAction:" destination="cEo-Qi-bYI" eventType="touchUpInside" id="uju-Ag-uuS"/>
                                         </connections>
                                     </button>
                                     <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="8tK-Tt-zyE">
-                                        <rect key="frame" x="12" y="12" width="120" height="18"/>
+                                        <rect key="frame" x="12" y="12" width="120" height="24"/>
                                         <constraints>
                                             <constraint firstAttribute="width" constant="120" id="TBT-4V-5It"/>
                                         </constraints>
-                                        <fontDescription key="fontDescription" name="BarlowSemiCondensed-Medium" family="Barlow Semi Condensed Medium" pointSize="15"/>
+                                        <fontDescription key="fontDescription" name="BarlowSemiCondensed-Medium" family="Barlow Semi Condensed Medium" pointSize="20"/>
                                         <color key="textColor" name="brainwalletContent"/>
                                         <color key="highlightedColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                                     </label>
                                 </subviews>
                                 <color key="backgroundColor" systemColor="systemPinkColor"/>
                                 <constraints>
+                                    <constraint firstItem="jLr-H0-dij" firstAttribute="top" secondItem="8fw-XB-zyN" secondAttribute="top" constant="12" id="2nj-w7-WHK"/>
                                     <constraint firstAttribute="height" constant="80" id="8em-HC-37f"/>
                                     <constraint firstItem="8tK-Tt-zyE" firstAttribute="top" secondItem="8fw-XB-zyN" secondAttribute="top" constant="12" id="I0f-mL-tM8"/>
                                     <constraint firstItem="8tK-Tt-zyE" firstAttribute="leading" secondItem="8fw-XB-zyN" secondAttribute="leading" constant="12" id="NGn-Kg-thI"/>
+                                    <constraint firstAttribute="trailing" secondItem="jLr-H0-dij" secondAttribute="trailing" constant="12" id="pVC-kQ-WwZ"/>
                                 </constraints>
                             </view>
                             <containerView opaque="NO" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="vHe-dR-fDI">
@@ -152,7 +159,7 @@
             <color red="0.72156862745098038" green="0.72156862745098038" blue="0.72156862745098038" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
         </namedColor>
         <systemColor name="systemPinkColor">
-            <color red="1" green="0.1764705882" blue="0.33333333329999998" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+            <color red="1" green="0.17647058823529413" blue="0.33333333333333331" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
         </systemColor>
     </resources>
 </document>
